+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/IndentationRule.kt
@@ -25,6 +25,7 @@ import com.pinterest.ktlint.core.ast.ElementType.ELSE
 import com.pinterest.ktlint.core.ast.ElementType.ELVIS
 import com.pinterest.ktlint.core.ast.ElementType.EOL_COMMENT
 import com.pinterest.ktlint.core.ast.ElementType.EQ
+import com.pinterest.ktlint.core.ast.ElementType.FOR
 import com.pinterest.ktlint.core.ast.ElementType.FUN
 import com.pinterest.ktlint.core.ast.ElementType.FUNCTION_LITERAL
 import com.pinterest.ktlint.core.ast.ElementType.GT
@@ -43,6 +44,7 @@ import com.pinterest.ktlint.core.ast.ElementType.OPEN_QUOTE
 import com.pinterest.ktlint.core.ast.ElementType.OPERATION_REFERENCE
 import com.pinterest.ktlint.core.ast.ElementType.PARENTHESIZED
 import com.pinterest.ktlint.core.ast.ElementType.PROPERTY_ACCESSOR
+import com.pinterest.ktlint.core.ast.ElementType.PROPERTY_DELEGATE
 import com.pinterest.ktlint.core.ast.ElementType.RBRACE
 import com.pinterest.ktlint.core.ast.ElementType.RBRACKET
 import com.pinterest.ktlint.core.ast.ElementType.REGULAR_STRING_PART
@@ -203,11 +205,11 @@ public class IndentationRule :
         autoCorrect: Boolean,
         emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
     ) {
+        val ctx = IndentContext()
         val firstNotEmptyLeaf = node.nextLeaf()
         if (firstNotEmptyLeaf?.let { it.elementType == WHITE_SPACE && !it.textContains('\n') } == true) {
-            visitWhiteSpace(firstNotEmptyLeaf, autoCorrect, emit)
+            visitWhiteSpace(firstNotEmptyLeaf, autoCorrect, emit, ctx)
         }
-        val ctx = IndentContext()
         node.visit(
             { n ->
                 when (n.elementType) {
@@ -243,11 +245,18 @@ public class IndentationRule :
                                 )
                             }
                             else -> {
-                                expectedIndent++
-                                logger.trace { "$line: block starting with ${n.text} -> Increase to $expectedIndent" }
-                                ctx.blockStack.push(
-                                    Block(n.elementType, line, REGULAR)
-                                )
+                                if (n.isPartOfForLoopConditionWithMultilineExpression()) {
+                                    logger.trace { "$line: block starting with ${n.text} -> Keep at $expectedIndent" }
+                                    ctx.blockStack.push(
+                                        Block(n.elementType, line, SAME_AS_PREVIOUS_BLOCK)
+                                    )
+                                } else {
+                                    expectedIndent++
+                                    logger.trace { "$line: block starting with ${n.text} -> Increase to $expectedIndent" }
+                                    ctx.blockStack.push(
+                                        Block(n.elementType, line, REGULAR)
+                                    )
+                                }
                             }
                         }
                         logger.trace {
@@ -400,14 +409,14 @@ public class IndentationRule :
                                         // )
                                         adjustExpectedIndentAfterLparInsideCondition(n, ctx)
                                 }
-                                visitWhiteSpace(n, autoCorrect, emit)
+                                visitWhiteSpace(n, autoCorrect, emit, ctx)
                                 if (ctx.localAdj != 0) {
                                     expectedIndent += ctx.localAdj
                                     logger.trace { "$line: ++${ctx.localAdj} on whitespace containing new line (${n.elementType}) -> $expectedIndent" }
                                     ctx.localAdj = 0
                                 }
                             } else if (n.isPartOf(KDOC)) {
-                                visitWhiteSpace(n, autoCorrect, emit)
+                                visitWhiteSpace(n, autoCorrect, emit, ctx)
                             }
                             line += n.text.count { it == '\n' }
                         }
@@ -439,10 +448,17 @@ public class IndentationRule :
         }) ?: return
         val nextSibling = n.treeNext
         if (!ctx.ignored.contains(p) && nextSibling != null) {
-            expectedIndent++
-            logger.trace { "$line: ++inside(${p.elementType}) -> $expectedIndent" }
-            ctx.ignored.add(p)
-            ctx.exitAdjBy(p, -1)
+            if (p.treeParent.elementType == PROPERTY_DELEGATE) {
+                expectedIndent += 2
+                logger.trace { "$line: ++dot-qualified-expression in property delegate -> $expectedIndent" }
+                ctx.ignored.add(p)
+                ctx.exitAdjBy(p, -2)
+            } else {
+                expectedIndent++
+                logger.trace { "$line: ++inside(${p.elementType}) -> $expectedIndent" }
+                ctx.ignored.add(p)
+                ctx.exitAdjBy(p, -1)
+            }
         }
     }
 
@@ -720,7 +736,8 @@ public class IndentationRule :
     private fun visitWhiteSpace(
         node: ASTNode,
         autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
+        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit,
+        ctx: IndentContext
     ) {
         val text = node.text
         val nodeIndent = text.substringAfterLast("\n")
@@ -770,7 +787,12 @@ public class IndentationRule :
             nextLeafElementType == GT &&
                 node.treeParent?.elementType.let { it == TYPE_PARAMETER_LIST || it == TYPE_ARGUMENT_LIST } ->
                 0
-            nextLeafElementType in rTokenSet -> -1
+            nextLeafElementType in rTokenSet ->
+                if (node.isPartOfForLoopConditionWithMultilineExpression()) {
+                    0
+                } else {
+                    -1
+                }
             // IDEA quirk:
             // val i: Int
             //     by lazy { 1 }
@@ -801,7 +823,30 @@ public class IndentationRule :
             //  var value: DataClass =
             //      DataClass("too long line")
             //          private set
-            node.nextCodeSibling()?.elementType == PROPERTY_ACCESSOR && node.treeParent.findChildByType(EQ)?.nextLeaf().isWhiteSpaceWithNewline() -> -1
+            node.nextCodeSibling()?.elementType == PROPERTY_ACCESSOR &&
+                node.treeParent.findChildByType(EQ)?.nextLeaf().isWhiteSpaceWithNewline() -> {
+                // Fix the indent of the body block/expression of the setter/getter
+                //  IDEA formatting:
+                //    private var foo: String =
+                //        "foo"
+                //        get() = value
+                //        set(value) {
+                //            field = value
+                //        }
+                //  instead of
+                //    private var foo: String =
+                //        "foo"
+                //        get() = value
+                //        set(value) {
+                //                field = value
+                //            }
+                expectedIndent--
+                val propertyAccessor = node.nextCodeSibling()!!
+                ctx.exitAdjBy(propertyAccessor, 1)
+
+                // Fix the indent before the setter/getter
+                -1
+            }
             else -> 0
         }
         // indentation with incorrect characters replaced
@@ -1037,3 +1082,32 @@ private fun KtStringTemplateExpression.isFollowedByTrimMargin() = isFollowedBy("
 private fun KtStringTemplateExpression.isFollowedBy(callExpressionName: String) =
     this.node.nextSibling { it.elementType != DOT }
         .let { it?.elementType == CALL_EXPRESSION && it.text == callExpressionName }
+
+/**
+ *  A for-loop for which the condition contains a sibling node containing a newline is not correctly formatted by the
+ *  default formatter of IntelliJ IDEA (https://youtrack.jetbrains.com/issue/IDEA-293691/Format-Kotlin-for-loop). When
+ *  using the correct indentation level, it conflicts with the IntelliJ IDEA formatting, so until the aforementioned bug
+ *  is resolved, ktlint will produce the same format as IntelliJ default formatter.
+ */
+private fun ASTNode.isPartOfForLoopConditionWithMultilineExpression(): Boolean {
+    if (treeParent.elementType != FOR) {
+        return false
+    }
+    if (this.elementType != LPAR) {
+        return treeParent.findChildByType(LPAR)!!.isPartOfForLoopConditionWithMultilineExpression()
+    }
+    require(elementType == LPAR) {
+        "Node should be the LPAR of the FOR loop"
+    }
+
+    // Iterate all sibling node until RPAR to check whether the node contains a newline. Note that it does not matter
+    // whether is code sibling contains a newline.
+    var node: ASTNode? = this
+    while (node != null && node.elementType != RPAR) {
+        if (node.isWhiteSpaceWithNewline()) {
+            return true
+        }
+        node = node.nextSibling { true }
+    }
+    return false
+}