+++ CHANGELOG.md
@@ -26,9 +26,14 @@ An AssertJ style API for testing KtLint rules ([#1444](https://github.com/pinter
 - Fix check of spacing in the receiver type of an anonymous function ([#1440](https://github.com/pinterest/ktlint/issues/1440))
 - Allow comment on same line as super class in class declaration `wrapping` ([#1457](https://github.com/pinterest/ktlint/pull/1457))
 - Respect git hooksPath setting ([#1465](https://github.com/pinterest/ktlint/issues/1465))
+- Fix formatting of a property delegate with a dot-qualified-expression `indent` ([#1340](https://github.com/pinterest/ktlint/ssues/1340))
+- Keep formatting of for-loop in sync with default IntelliJ formatter (`indent`) and a newline in the expression in a for-statement should not force to wrap it `wrapping` ([#1350](https://github.com/pinterest/ktlint/issues/1350))
+- Fix indentation of property getter/setter when the property has an initializer on a separate line `indent` ([#1335](https://github.com/pinterest/ktlint/issues/1335))
 
 ### Changed
-* Set Kotlin development version to `1.7.0-Beta` and Kotlin version to `1.6.21`.
+- Update Kotlin development version to `1.7.0-RC` and Kotlin version to `1.6.21`.
+- Update shadow plugin to `7.1.2` release
+- Update picocli to `4.6.3` release
 
 ### Removed